Fix footer, fix some auth and create post details page

+++ app/Http/Controllers/PostController.php
@@ -30,6 +30,10 @@ class PostController extends Controller
         return redirect()->route('posts.index');
     }
 
+    public function show(Post $post){
+        return view('posts.show', compact('post'));
+    }
+
     public function edit(Post $post){
         if (auth()->id() !== $post->user_id) {
             abort(403);
@@ -38,6 +42,10 @@ class PostController extends Controller
     }
 
     public function update(Request $request, Post $post){
+        if (auth()->id() !== $post->user_id) {
+            abort(403);
+        }
+
         $request->validate([
             'title' => 'required',
             'content' => 'required',
@@ -49,6 +57,9 @@ class PostController extends Controller
     }
 
     public function destroy(Post $post){
+        if (auth()->id() !== $post->user_id) {
+            abort(403);
+        }
         $post->delete();
 
         return redirect()->route('posts.index');
